Split game from engine directories
