+++ client/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Switch } from "react-router-dom";
 import React from "react";
 import Login from "./routes/Login";
 import Home from "./routes/Home";
-import Wiz from "./routes/Wiz";
+import Wiz from "./wiz/routes/Wiz";
 import Room from "./routes/Room";
 import Lobby from "./routes/Lobby";
 import Rooms from "./routes/RoomCreate";
