+++ client/src/engine/GameSelector.ts
@@ -1,4 +1,4 @@
-import { WizApi } from "../api/WizApi";
+import { WizApi } from "../wiz/api/WizApi";
 
 export default function createGame(name: string) {
     switch (name.toLowerCase()){
