+++ client/src/wiz/components/Bet.tsx
