+++ client/src/wiz/components/CardFan.tsx
