+++ client/src/wiz/components/EmptyCard.tsx
