+++ client/src/wiz/components/EmptyCardFan.tsx
