+++ client/src/wiz/components/OtherPlayer.tsx
