+++ client/src/wiz/components/OtherPlayers.tsx
