+++ client/src/wiz/components/PlaceholderCard.tsx
