+++ client/src/wiz/components/PlayerList.tsx
