+++ client/src/wiz/components/PlayerListItem.tsx
