+++ client/src/wiz/components/ScoreBoard.tsx
