+++ client/src/wiz/components/StrongSuit.tsx
