+++ client/src/wiz/interfaces/PossibleMoves.ts
